requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
astor==0.8.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmaes==0.9.1
cmake==3.26.3
colorama==0.4.6
colorlog==6.7.0
contextlib2==21.6.0
contourpy==1.0.7
cucim==23.4.1
cycler==0.11.0
databricks-cli==0.17.7
docker==6.1.2
einops==0.6.1
entrypoints==0.4
filelock==3.12.0
fire==0.5.0
Flask==2.3.2
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.54.0
gunicorn==20.1.0
h5py==3.8.0
huggingface-hub==0.14.1
idna==3.4
imagecodecs==2023.3.16
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
itk==5.3.0
itk-core==5.3.0
itk-filtering==5.3.0
itk-io==5.3.0
itk-numerics==5.3.0
itk-registration==5.3.0
itk-segmentation==5.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
json-tricks==3.16.1
jsonschema==4.17.3
kiwisolver==1.4.4
lazy-loader==0.2
lightning-utilities==0.8.0
lit==16.0.5
lmdb==1.4.1
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
miacag==0.0.81
mlflow==2.3.1
monai==1.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nibabel==5.1.0
ninja==1.11.1
nni==2.10
nptyping==2.5.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
openslide-python==1.1.2
optuna==3.1.1
packaging==23.1
pandas==1.2.4
patsy==0.5.3
Pillow==9.5.0
prettytable==3.7.0
protobuf==3.20.3
psutil==5.9.5
psycopg2==2.9.6
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydicom==2.3.1
PyJWT==2.7.0
pynrrd==1.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
PythonWebHDFS==0.2.3
pytorch-ignite==0.4.10
pytorch-lightning==2.0.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
querystring-parser==1.2.4
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
responses==0.23.1
rsa==4.9
schema==0.7.5
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
SimpleITK==2.2.1
simplejson==3.19.1
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.13
sqlparse==0.4.4
statsmodels==0.14.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboardX==2.6
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tokenizers==0.12.1
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
transformers==4.21.3
triton==2.0.0
typeguard==4.0.0
types-PyYAML==6.0.12.9
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
